+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- When `kwargs.json` is manually filled with `null`, it will be refreshed under the assumption
+- When `kwargs.json` is manually filled with `null`, it is refreshed under the assumption
   that the existing results are consistent with the new `kwargs.json` file.
   This is useful when one is refactoring workflows and one wants to reuse some results of
   a previous run.
+- When `kwargs.sha256` is manually removed, it is refreshed for the same reasons as in the
+  previous point.
 
 ### Fixed