+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix mistake in computation of SHA256 hash: now it reads the entire file.
+
 ### Added
 
 - New utility script to remove unfinished jobs from a results directory: `parman-clean-results`.